requirements.txt
aiohttp==3.8.3
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.1.3
djangorestframework==3.14.0
bcrypt==4.0.1
requests==2.28.1
bs4~=0.0.1
psycopg2
python-dotenv